[feature-message] 메시지 읽음 개수 가져오는 기능 테스트코드 작성

+++ src/main/java/com/flextalk/we/message/dto/MessageReadDto.java
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 public class MessageReadDto {
 
     private Long messageId;
-    private int messageReadCount;
+    private Long messageReadCount;
 
-    public MessageReadDto(Long messageId, int messageReadCount) {
+    public MessageReadDto(Long messageId, Long messageReadCount) {
         this.messageId = messageId;
         this.messageReadCount = messageReadCount;
     }
